Modify separator symbols

+++ scripts/tmux.sh
@@ -1,7 +1,50 @@
 #!/bin/bash
+
+# options:
+help_msg="
+options:
+    -h, --help: show help messages
+    -a, --separator-arrow: use arrow symbols as separators
+"
+
+separator_arrow=false
+is_help=false
+for i in $*; do
+    if [ $i == "-a" ] || [ $i == "--separator-arrow" ]
+    then
+        separator_arrow=true
+    elif [ $i == "-h" ] || [ $i == "--help" ]
+    then
+        is_help=true
+    fi
+done
+
+
+if [ $is_help == true ]
+then
+    echo "${help_msg}"
+    exit 0
+fi
+
+
 rm -f ${HOME}/.tmux.conf
 rm -f ${HOME}/.tmux.conf.local
 
 sudo apt-get install tmux autojump -y
 ln -s -f $(pwd)/.tmux/.tmux.conf ${HOME}/.tmux.conf
 cp $(pwd)/.tmux/.tmux.conf.local ${HOME}/.tmux.conf.local
+
+
+# modify separator symbols
+if [ $separator_arrow == true ]
+then
+    sed -i 's/\(tmux_conf_theme_left_separator_main="\).*\("\)/\1\\uE0B0\2/g' ~/.tmux.conf.local
+    sed -i 's/\(tmux_conf_theme_left_separator_sub="\).*\("\)/\1\\uE0B1\2/g' ~/.tmux.conf.local
+    sed -i 's/\(tmux_conf_theme_right_separator_main="\).*\("\)/\1\\uE0B2\2/g' ~/.tmux.conf.local
+    sed -i 's/\(tmux_conf_theme_right_separator_sub="\).*\("\)/\1\\uE0B3\2/g' ~/.tmux.conf.local
+else
+    sed -i 's/\(tmux_conf_theme_left_separator_main="\).*\("\)/\1\2/g' ~/.tmux.conf.local
+    sed -i 's/\(tmux_conf_theme_left_separator_sub="\).*\("\)/\1|\2/g' ~/.tmux.conf.local
+    sed -i 's/\(tmux_conf_theme_right_separator_main="\).*\("\)/\1\2/g' ~/.tmux.conf.local
+    sed -i 's/\(tmux_conf_theme_right_separator_sub="\).*\("\)/\1|\2/g' ~/.tmux.conf.local
+fi
